fix for large DeviceMedia (when over 2 GB) - DWORD changed to int64, lseek64 used
git-svn-id: https://subversion.assembla.com/svn/atarijookie/trunk@658 0d2ca917-fb58-4453-bb26-a698c01b0940

+++ cosmosex/native/imagefilemedia.cpp
@@ -89,13 +89,13 @@ void ImageFileMedia::setMediaChanged(bool changed)
     mediaHasChanged = changed;
 }
 
-void ImageFileMedia::getCapacity(DWORD &bytes, DWORD &sectors)
+void ImageFileMedia::getCapacity(int64_t &bytes, int64_t &sectors)
 {
     bytes   = BCapacity;
     sectors = SCapacity;
 }
 
-bool ImageFileMedia::readSectors(DWORD sectorNo, DWORD count, BYTE *bfr)
+bool ImageFileMedia::readSectors(int64_t sectorNo, DWORD count, BYTE *bfr)
 {
     if(!isInit()) {                             // if not initialized, failed
         return false;
@@ -118,7 +118,7 @@ bool ImageFileMedia::readSectors(DWORD sectorNo, DWORD count, BYTE *bfr)
     return true;
 }
 
-bool ImageFileMedia::writeSectors(DWORD sectorNo, DWORD count, BYTE *bfr)
+bool ImageFileMedia::writeSectors(int64_t sectorNo, DWORD count, BYTE *bfr)
 {
     if(!isInit()) {                             // if not initialized, failed
         return false;
